web-app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
